backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.2
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.3
python-multipart>=0.0.9
openai>=1.30.1
python-dotenv>=1.0.1
google-genai>=0.1.0
pydantic[email]>=2.7.1
httpx>=0.27.0
pydantic-settings>=2.3.1
mongomock-motor>=0.0.34
